Split Apdb interface into two classes.
Replication-related methods now appear in a separate class `ApdbReplica`.
Instances of `ApdbReplica` are made from the same ApdbConfig.
