+++ python/lsst/dax/apdb/apdbSql.py
@@ -42,10 +42,12 @@ from lsst.utils.iteration import chunk_iterable
 from sqlalchemy import func, sql
 from sqlalchemy.pool import NullPool
 
-from .apdb import Apdb, ApdbConfig, ApdbInsertId, ApdbTableData
+from .apdb import Apdb, ApdbConfig
 from .apdbConfigFreezer import ApdbConfigFreezer
 from .apdbMetadataSql import ApdbMetadataSql
+from .apdbReplica import ApdbInsertId
 from .apdbSchema import ApdbTables
+from .apdbSqlReplica import ApdbSqlReplica
 from .apdbSqlSchema import ApdbSqlSchema, ExtraTables
 from .timer import Timer
 from .versionTuple import IncompatibleVersionError, VersionTuple
@@ -58,8 +60,15 @@ if TYPE_CHECKING:
 _LOG = logging.getLogger(__name__)
 
 VERSION = VersionTuple(0, 1, 0)
-"""Version for the code defined in this module. This needs to be updated
-(following compatibility rules) when schema produced by this code changes.
+"""Version for the code controlling non-replication tables. This needs to be
+updated following compatibility rules when schema produced by this code
+changes.
+"""
+
+REPLICA_VERSION = VersionTuple(0, 1, 0)
+"""Version for the code controlling replication tables. This needs to be
+updated following compatibility rules when schema produced by this code
+changes.
 """
 
 
@@ -166,20 +175,6 @@ class ApdbSqlConfig(ApdbConfig):
         super().validate()
         if len(self.ra_dec_columns) != 2:
             raise ValueError("ra_dec_columns must have exactly two column names")
-
-
-class ApdbSqlTableData(ApdbTableData):
-    """Implementation of ApdbTableData that wraps sqlalchemy Result."""
-
-    def __init__(self, result: sqlalchemy.engine.Result):
-        self._keys = list(result.keys())
-        self._rows: list[tuple] = cast(list[tuple], list(result.fetchall()))
-
-    def column_names(self) -> list[str]:
-        return self._keys
-
-    def rows(self) -> Iterable[tuple]:
-        return self._rows
 
 
 class ApdbSql(Apdb):
@@ -425,6 +420,10 @@ class ApdbSql(Apdb):
         # Docstring inherited from base class.
         return self._schema.schemaVersion()
 
+    def get_replica(self) -> ApdbSqlReplica:
+        """Return `ApdbReplica` instance for this database."""
+        return ApdbSqlReplica(self._schema, self._engine)
+
     def tableRowCount(self) -> dict[str, int]:
         """Return dictionary with the table names and row counts.
 
@@ -597,78 +596,6 @@ class ApdbSql(Apdb):
             else:
                 result = conn.execute(query2).scalar_one_or_none()
                 return result is not None
-
-    def getInsertIds(self) -> list[ApdbInsertId] | None:
-        # docstring is inherited from a base class
-        if not self._schema.has_insert_id:
-            return None
-
-        table = self._schema.get_table(ExtraTables.DiaInsertId)
-        assert table is not None, "has_insert_id=True means it must be defined"
-        query = sql.select(table.columns["insert_id"], table.columns["insert_time"]).order_by(
-            table.columns["insert_time"]
-        )
-        with Timer("DiaObject insert id select", self.config.timer):
-            with self._engine.connect() as conn:
-                result = conn.execution_options(stream_results=True, max_row_buffer=10000).execute(query)
-                ids = []
-                for row in result:
-                    insert_time = astropy.time.Time(row[1].timestamp(), format="unix_tai")
-                    ids.append(ApdbInsertId(id=row[0], insert_time=insert_time))
-                return ids
-
-    def deleteInsertIds(self, ids: Iterable[ApdbInsertId]) -> None:
-        # docstring is inherited from a base class
-        if not self._schema.has_insert_id:
-            raise ValueError("APDB is not configured for history storage")
-
-        table = self._schema.get_table(ExtraTables.DiaInsertId)
-
-        insert_ids = [id.id for id in ids]
-        where_clause = table.columns["insert_id"].in_(insert_ids)
-        stmt = table.delete().where(where_clause)
-        with self._engine.begin() as conn:
-            conn.execute(stmt)
-
-    def getDiaObjectsHistory(self, ids: Iterable[ApdbInsertId]) -> ApdbTableData:
-        # docstring is inherited from a base class
-        return self._get_history(ids, ApdbTables.DiaObject, ExtraTables.DiaObjectInsertId)
-
-    def getDiaSourcesHistory(self, ids: Iterable[ApdbInsertId]) -> ApdbTableData:
-        # docstring is inherited from a base class
-        return self._get_history(ids, ApdbTables.DiaSource, ExtraTables.DiaSourceInsertId)
-
-    def getDiaForcedSourcesHistory(self, ids: Iterable[ApdbInsertId]) -> ApdbTableData:
-        # docstring is inherited from a base class
-        return self._get_history(ids, ApdbTables.DiaForcedSource, ExtraTables.DiaForcedSourceInsertId)
-
-    def _get_history(
-        self,
-        ids: Iterable[ApdbInsertId],
-        table_enum: ApdbTables,
-        history_table_enum: ExtraTables,
-    ) -> ApdbTableData:
-        """Return catalog of records for given insert identifiers, common
-        implementation for all DIA tables.
-        """
-        if not self._schema.has_insert_id:
-            raise ValueError("APDB is not configured for history retrieval")
-
-        table = self._schema.get_table(table_enum)
-        history_table = self._schema.get_table(history_table_enum)
-
-        join = table.join(history_table)
-        insert_ids = [id.id for id in ids]
-        history_id_column = history_table.columns["insert_id"]
-        apdb_columns = self._schema.get_apdb_columns(table_enum)
-        where_clause = history_id_column.in_(insert_ids)
-        query = sql.select(history_id_column, *apdb_columns).select_from(join).where(where_clause)
-
-        # execute select
-        with Timer(f"{table.name} history select", self.config.timer):
-            with self._engine.begin() as conn:
-                result = conn.execution_options(stream_results=True, max_row_buffer=10000).execute(query)
-                return ApdbSqlTableData(result)
 
     def getSSObjects(self) -> pandas.DataFrame:
         # docstring is inherited from a base class
