+++ python/lsst/dax/apdb/factory.py
@@ -28,6 +28,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from .apdb import Apdb, ApdbConfig
     from .apdbCassandra import ApdbCassandra
+    from .apdbReplica import ApdbReplica
     from .apdbSql import ApdbSql
 
 
@@ -84,4 +85,32 @@ def make_apdb(config: ApdbConfig) -> Apdb:
         return ApdbSql(config)
     elif type(config) is ApdbCassandraConfig:
         return ApdbCassandra(config)
+    raise TypeError(f"Unknown type of config object: {type(config)}")
+
+
+def make_apdb_replica(config: ApdbConfig) -> ApdbReplica:
+    """Create ApdbReplica instance based on Apdb configuration.
+
+    Parameters
+    ----------
+    config : `ApdbConfig`
+        Configuration object, sub-class of ApdbConfig.
+
+    Returns
+    -------
+    apdb_replica : `ApdbReplica`
+        Instance of a specific ApdbReplica sub-class.
+
+    Raises
+    ------
+    TypeError
+        Raised if type of ``config`` does not match any known types.
+    """
+    from .apdbCassandra import ApdbCassandra, ApdbCassandraConfig
+    from .apdbSql import ApdbSql, ApdbSqlConfig
+
+    if type(config) is ApdbSqlConfig:
+        return ApdbSql(config).get_replica()
+    elif type(config) is ApdbCassandraConfig:
+        return ApdbCassandra(config).get_replica()
     raise TypeError(f"Unknown type of config object: {type(config)}")
